New ETC/ETH deposit server script
It doesn't require smart contracts to operate. Binary search is used to
detect deposits in a range of blocks. Unfortunately, it doesn't work for
the whole range of blocks from the block 0. History depth for ropsten
network is about 2 months, and about 1 month for mainnet. Deposits made
by smartcontracts are not detected correctly. Instead, we can only
detect the block which could bring several deposits to the single
address. These several deposits won't be detected separetely. Client
software will be notified about block hash in this case.

Deposits of 0 ETH are not detected which doesn't seem to be a problem.